# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Headway API",
    "description": "Serves computed transit metrics with full lineage, the DQ resolution workflow, and the audited certification action. This API never computes a reported figure; it serves what the calculation library produced, joined to its provenance. Reported values are JSON strings (exact NUMERIC, never float). Auth: local accounts (ADR-0011); the native OIDC relying party is the next increment behind the same {sub, username, role} claim set.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sandbox/preview": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Sandbox Preview",
        "description": "Preview the figure impact of proposed policy-knob values for one\nperiod \u2014 read-only, ephemeral, never certifiable (module docstring).\nBaseline = the CURRENT audited settings (app.settings, the same\nprecedence a real run uses); proposed = the same computation with only\nthe proposed knobs overridden.",
        "operationId": "sandbox_preview_sandbox_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewMetricImpact": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "calc_name": {
            "type": "string",
            "title": "Calc Name"
          },
          "calc_version": {
            "type": "string",
            "title": "Calc Version"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "baseline": {
            "$ref": "#/components/schemas/PreviewSide"
          },
          "proposed": {
            "$ref": "#/components/schemas/PreviewSide"
          },
          "delta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "calc_name",
          "calc_version",
          "unit",
          "scope",
          "category",
          "baseline",
          "proposed",
          "delta"
        ],
        "title": "PreviewMetricImpact"
      },
      "PreviewSection": {
        "properties": {
          "baseline_thresholds": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Baseline Thresholds"
          },
          "baseline_threshold_sources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Baseline Threshold Sources"
          },
          "proposed_thresholds": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Proposed Thresholds"
          },
          "inputs": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Inputs"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/PreviewMetricImpact"
            },
            "type": "array",
            "title": "Metrics"
          },
          "derivation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derivation"
          }
        },
        "type": "object",
        "required": [
          "baseline_thresholds",
          "baseline_threshold_sources",
          "proposed_thresholds",
          "inputs",
          "metrics"
        ],
        "title": "PreviewSection"
      },
      "PreviewSide": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "blocked": {
            "type": "boolean",
            "title": "Blocked"
          },
          "findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Findings"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "value",
          "blocked",
          "findings",
          "detail"
        ],
        "title": "PreviewSide",
        "description": "One variant's outcome for one metric: the would-be value (or an\nhonest refusal) plus every would-be finding. Nothing here was written\nanywhere."
      },
...
      "SandboxPreviewRequest": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "proposed": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "minProperties": 1,
            "title": "Proposed"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "proposed"
        ],
        "title": "SandboxPreviewRequest"
      },
      "SandboxPreviewResponse": {
        "properties": {
          "banner": {
            "type": "string",
            "title": "Banner"
          },
          "persisted": {
            "type": "boolean",
            "title": "Persisted"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "period_convention": {
            "type": "string",
            "title": "Period Convention"
          },
          "proposed": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Proposed"
          },
          "settings_flow_note": {
            "type": "string",
            "title": "Settings Flow Note"
          },
          "ntd": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviewSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "ops": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviewSection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "banner",
          "persisted",
          "period_start",
          "period_end",
          "period_convention",
          "proposed",
          "settings_flow_note"
        ],
        "title": "SandboxPreviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}